evently/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
python-dotenv==1.0.1
httpx==0.27.2
bcrypt==3.2.2
redis>=5.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=6.1.0
python-json-logger>=2.0.7



